+++ text_authorship/ta_model/__init__.py
@@ -1,3 +1,3 @@
 from .data_extraction import load_df
 from .model_selection import train_crossval_twofold, train_test_split, get_encoders
-from .data_preparation import FeatureBuilder
+from featurebuilder import FeatureBuilder
