+++ text_authorship/ta_model/data_preparation.py
@@ -1,236 +1,11 @@
-import pandas as pd
-
-import sys
-import unicodedata
-import nltk
-from nltk import word_tokenize
-from nltk.corpus import stopwords
-from pymorphy2 import MorphAnalyzer
-import numpy as np
 from itertools import chain
-import scipy as sp
-from enum import Flag, auto
-from collections import namedtuple
 
-from sklearn.base import BaseEstimator, TransformerMixin
+import pandas as pd
+import nltk
+import numpy as np
+
 from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.preprocessing import LabelEncoder
-
-
-PUNKT = {chr(i) for i in range(sys.maxunicode) if unicodedata.category(chr(i)).startswith('P')}
-
-
-def _strip_str(s: str, chars):
-    i, j = 0, len(s)
-
-    while i < j and s[i] in chars:
-        i += 1
-
-    while i < j and s[j - 1] in chars:
-        j -= 1
-
-    return s[i:j]
-
-
-nltk.download('stopwords')
-
-
-STOP_TAGS = [
-    'Abbr',
-    'Name',
-    'Surn',
-    'Patr',
-    'Geox',
-    'Orgn',
-    'Trad'
-]
-
-DELETED = 'deleted'
-
-
-class TATransformer(BaseEstimator, TransformerMixin):
-
-    _parsers = []
-
-    def __init__(self, use_stopwords=True, save_path='transformed_df.csv', load_path=None):
-        self.use_stopwords = use_stopwords
-        self.save_path = save_path
-        self.load_path = load_path
-    
-    def fit(self, X: pd.DataFrame, y=None):
-        self.morph_ = MorphAnalyzer()
-        self.stopwords_ = set()
-        self.new_cols_ = ParseManager.get_col_names()
-
-        if self.use_stopwords:
-            self.stopwords_ = self.stopwords_ | set(stopwords.words('russian'))
-
-        return self
-    
-    def transform(self, X: pd.DataFrame):
-        if self.load_path and self.load_path.endswith('.csv'):
-            X = pd.read_csv(self.load_path)
-        else:
-            X = X.copy()
-            X['text'] = X['text'].str.lower()
-
-            new_data = pd.DataFrame(
-                list(map(self.parse_text, X['text'])),
-                columns=self.new_cols_
-            )
-
-            X = pd.concat([
-                X,
-                new_data
-            ], axis=1)
-
-            if self.save_path:
-                if not self.save_path.endswith('.csv'):
-                    raise ValueError("expected csv destination file")
-                X.to_csv(self.save_path, index=False)
-
-        return X
-    
-    def parse_text(self, text: str):
-        return ParseManager.parse_text(self.morph_, self.stopwords_, text)
-    
-
-class TokenType(Flag):
-    NOFLAG = 0
-    PUNKT = auto()
-    STOPWORD = auto()
-    STOPTAG = auto()
-
-
-ParsingParams = namedtuple('ParsingParams', 'token stripped anls token_type')
-
-
-class ParseManager:
-    parsers = []
-
-    @classmethod
-    def register_parser(cls, parser):
-        cls.parsers.append(parser)
-        return parser
-    
-    @classmethod
-    def get_col_names(cls):
-        cols = [parser.col_name for parser in cls.parsers]
-        return cols
-    
-    @classmethod
-    def parse_text(cls, morph: MorphAnalyzer, sw, text: str):
-        tools = [P() for P in cls.parsers]
-
-        for token in word_tokenize(text, language='russian'):
-            token_type = TokenType.NOFLAG
-            stripped = _strip_str(token, PUNKT)
-
-            if not stripped:
-                token_type = token_type | TokenType.PUNKT
-
-            anls = morph.parse(stripped)[0]
-
-            if anls.normal_form in sw:
-                token_type = token_type | TokenType.STOPWORD
-
-            if any([tag in anls.tag for tag in STOP_TAGS]):
-                token_type = token_type | TokenType.STOPTAG
-
-            params = ParsingParams(token=token,
-                                   stripped=stripped,
-                                   anls=anls,
-                                   token_type=token_type)
-
-            for p in tools:
-                p.parse_token(params)
-
-        result = [' '.join(p.tokens) for p in tools]
-        return result
-
-
-class BaseParser:
-
-    col_name = 'base'
-
-    def __init__(self):
-        self.tokens = []
-
-    def parse_token(self, params: ParsingParams):
-        raise NotImplementedError
-    
-
-@ParseManager.register_parser
-class WordParser(BaseParser):
-
-    col_name = 'text_no_punkt'
-
-    def __init__(self):
-        super().__init__()
-
-    def parse_token(self, params: ParsingParams):
-        if TokenType.PUNKT in params.token_type:
-            return
-        
-        if TokenType.STOPTAG in params.token_type:
-            self.tokens.append(DELETED)
-            return
-        
-        self.tokens.append(params.stripped)
-
-
-@ParseManager.register_parser
-class LemmaParser(BaseParser):
-
-    col_name = 'lemmas'
-
-    def __init__(self):
-        super().__init__()
-
-    def parse_token(self, params: ParsingParams):
-        if TokenType.PUNKT in params.token_type or TokenType.STOPWORD in params.token_type:
-            return
-        
-        if TokenType.STOPTAG in params.token_type:
-            self.tokens.append(DELETED)
-            return
-
-        self.tokens.append(params.anls.normal_form)
-
-
-@ParseManager.register_parser
-class TagParser(BaseParser):
-
-    col_name = 'tags'
-
-    def __init__(self):
-        super().__init__()
-
-    def parse_token(self, params: ParsingParams):
-        if TokenType.PUNKT in params.token_type or TokenType.STOPWORD in params.token_type:
-            return
-        
-        if TokenType.STOPTAG in params.token_type:
-            self.tokens.append(DELETED)
-            return
-
-        self.tokens.append(f'{len(params.stripped)}_{params.anls.tag.POS}')
-
-
-@ParseManager.register_parser
-class TokenParser(BaseParser):
-
-    col_name = 'tokens'
-
-    def __init__(self):
-        super().__init__()
-
-    def parse_token(self, params: ParsingParams):
-        if TokenType.STOPTAG in params.token_type:
-            self.tokens.append(DELETED)
-            return
-        
-        self.tokens.append(params.token)
 
 
 def get_document_vectorizer(frame, n_min=1, n=2, max_count=10000, column="lemmas"):
@@ -338,106 +113,3 @@ def check_seq(val):
     iterable = hasattr(val, "__iter__")
     not_string = not isinstance(val, str)
     return iterable and not_string
-
-
-class FeatureBuilder:
-    feature_mapping = {"tokens": "vectorizer",
-                       "text_no_punkt": "vectorizer",
-                       "lemmas": "vectorizer",
-                       "tags": "vectorizer"}
-
-    def __init__(self, *args, **vectorizers):
-        vectorizers = {k.lstrip("vec_"): v for k, v in vectorizers.items()}
-        featurelist = self.pack_features(args)
-        self.vectorizers = {}
-        for feature in featurelist:
-            processor = self.feature_mapping.get(feature, None)
-            if not processor:
-                self.vectorizers[feature] = None
-                continue
-            if processor != "vectorizer":
-                raise ValueError("only vectorizing is supported for non-scalar features")
-            vectorizer = vectorizers.get(feature, None)
-            if vectorizer is None:
-                raise ValueError(f"no vectorizer for feature: {feature}")
-            self.vectorizers[feature] = vectorizer
-        self.ordered_ft = list(sorted(featurelist,
-                                      key=lambda x: self.feature_mapping.get(x, "")))
-        self.ordered_proc = [self.feature_mapping.get(ft, None) for ft in self.ordered_ft]
-        self._initialized = False
-        self.feature_idx = None
-
-    @staticmethod
-    def pack_features(features):
-        attrs = (ft if check_seq(ft) else (ft,) for ft in features)
-        return list(chain(*attrs))
-
-    @staticmethod
-    def get_last_occurence(seq, val):
-        return len(seq) - 1 - seq[::-1].index(val)
-
-    @staticmethod
-    def get_first_occurence(seq, val):
-        return seq.index(val)
-
-    def _group_transform_features(self, df,  processor):
-        first_idx = self.get_first_occurence(self.ordered_proc, processor)
-        last_idx = self.get_last_occurence(self.ordered_proc, processor)
-        feature_slice = self.ordered_ft[first_idx:last_idx + 1]
-        feature_mat, positions = self.bulk_process(df, processor, feature_slice)
-        return feature_mat, positions
-
-    def fit_transform(self, df):
-        feature_positions = []
-        feature_matrices = []
-        for proc in set(self.ordered_proc):
-            featuremat, positions = self._group_transform_features(df, proc)
-            feature_matrices.extend(featuremat)
-            feature_positions.extend(positions)
-        final_matrix = sp.sparse.hstack(feature_matrices)
-        counter = 0
-        self.feature_idx = dict()
-        for ft, length in zip(self.ordered_ft, feature_positions):
-            self.feature_idx[ft] = (counter, length)
-            counter += length
-        self._initialized = True
-        return final_matrix
-
-    def transform(self, df):
-        feature_matrices = []
-        for proc in set(self.ordered_proc):
-            featuremat, positions = self._group_transform_features(df, proc)
-            feature_matrices.extend(featuremat)
-        final_matrix = sp.sparse.hstack(feature_matrices)
-        return final_matrix
-
-    def bulk_process(self, df, proc, featurelist):
-        if proc is None:
-            columns = df[featurelist].to_numpy(dtype=np.float64)
-            mat = [sp.sparse.csr_matrix(columns)]
-            indices = [1 for ft in featurelist]
-            return mat, indices
-        elif proc != "vectorizer":
-            raise ValueError("only vectorizers supported now")
-        else:
-            matrices = []
-            indices = []
-            for i, feature in enumerate(featurelist):
-                vectorizer = self.vectorizers[feature]
-                mat = vectorizer.transform(df[feature])
-                matrices.append(mat)
-                indices.append(mat.shape[1])
-            return matrices, indices
-
-    def find_idx(self, idx):
-        for key, (start, length) in self.feature_idx.items():
-            if start <= idx < (start + length):
-                break
-        else:
-            raise ValueError(f"index {idx} too big")
-        vectorizer = self.vectorizers[key]
-        if vectorizer is None:
-            return key
-        else:
-            features = vectorizer.get_feature_names_out()
-            return features[idx - start]
